feat: UserReportProfile 파일 분리
